refactor: extract duplicated components to shared locations
- Extract TagSelect to src/components/TagSelect/ (was duplicated x3)
- Extract StandardFormRow to src/components/StandardFormRow/ (was duplicated x3)
- Merge AvatarList into src/components/AvatarList/ (was duplicated x2)
- Merge ArticleListContent into src/components/ArticleListContent/ (was duplicated x2)
- Update all consumers to import from @/components

Removes ~600 lines of duplicated code across 10 component directories.
