+++ src/components/AvatarList/index.style.ts
