+++ src/components/index.ts
@@ -9,4 +9,12 @@ import Footer from './Footer';
 import { Question, SelectLang } from './RightContent';
 import { AvatarDropdown, AvatarName } from './RightContent/AvatarDropdown';
 
+/**
+ * 业务组件
+ */
+export { default as ArticleListContent } from './ArticleListContent';
+export { default as AvatarList } from './AvatarList';
+export { default as StandardFormRow } from './StandardFormRow';
+export { default as TagSelect } from './TagSelect';
+
 export { AvatarDropdown, AvatarName, Footer, Question, SelectLang };
