+++ src/pages/list/search/projects/index.tsx
@@ -4,10 +4,8 @@ import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
 import type { FC } from 'react';
 import { useState } from 'react';
+import { AvatarList, StandardFormRow, TagSelect } from '@/components';
 import { categoryOptions } from '../../mock';
-import AvatarList from './components/AvatarList';
-import StandardFormRow from './components/StandardFormRow';
-import TagSelect from './components/TagSelect';
 import type { ListItemDataType } from './data.d';
 import { queryFakeList } from './service';
 import useStyles from './style.style';
